Ensure sphinx compiles nice docs for seismology
