+++ lightkurve/seismology/__init__.py
@@ -1,2 +1,8 @@
-"""This sub-package provides classes and functions for asteroseismology."""
+"""The `lightkurve.seismology sub-package provides classes and functions for 
+quick-look asteroseismic analyses."""
 from .butler import *
+from .utils import *
+
+from .numax_estimators import *
+from .deltanu_estimators import *
+from .stellar_estimators import *
